Support new logging methods with legacy decorator
Add support for the new log_command_* methods with minimal changes to the
existing Formatter code. This is done by decorating the Command object to
make it behave like the legacy version that has non deprecated stderr/stdout 
methods. Since we override the @stdout/@stderr attributes, these must be 
filled with the data passed to log_command_data.

This commit is intended as a stopgap to restore compatibility with with
SSHKit @ master until more serious refactoring can be done.