Do not trigger command decoration side effects if log line is hidden
In cases where the log line is hidden (e.g. if it is a command at debug verbosity or command output is hidden), the decorate method shouldn’t be called because it has a side effect of incrementing the command number. This leads to ‘missing’ command numbers in the sequence in the console (issue #43). This commit fixes this issue.