Update tests to work with sshkit master / no user
This accounts for changes in sshkit's local username lookup introduced
by capistrano/sshkit#293 and also accounts for the conditional
formatting changes introduced by mattbrictson/airbrussh#66.

The latter did not previously cause failed tests, because the username
was stubbed in the tests, causing the relevant code to not be exercised.

Now the username is looked up through sshkit, which results in an empty
username on sshkit < 1.7.0, so the no-username display is triggered.

Unfortunately we need to keep the Etc stub for now to avoid crashing on
Windows with sshkit 1.7.0 and 1.7.1.