Only run failure hook if Airbrussh is being used
Since Airbrussh will be the default formatter of Capistrano, that means it will
always be loaded. However, a user may decide to use a different formatter, and
Airbrussh's deploy failure hook will still be registered. In this case the hook
should not be performed.

Solve this by asking SSHKit what the current formatter is, and skipping the
failure hook if the formatter isn't Airbrussh.