Allow Airbrussh to be used without rake installed
We only actually need rake if the `monkey_patch_rake` setting is true.
By default it is false, so airbrussh should be usable without rake.

However we were running `require "rake"` regardless of this config,
which meant that in practice you would get a `LoadError` if rake is
unavailable.

Fix by running `require "rake"` only if `monkey_patch_rake` is true.

Fixes #126