Ensure rubocop task runs on Ruby >= 3.1
We use an old version of rubocop (0.50.0) in order to support code
written for Ruby 1.9.3. This version of rubocop breaks on modern rubies
due to changes in the YAML API.

Installing the `safe_yaml` gem is a way to work around this. If rubocop
detects it, it will use the SafeYAML API instead of the one built into
Ruby.